Add integration tests for `Logger`
Also:
 * Unify test suite names.
